+++ nolimitlib/DependLibs/libplayer-nlc/xbmc/xbmc/filesystem/MusicDatabaseFile.h
@@ -22,7 +22,7 @@ public:
   bool Exists(const NlcUrl& url) override;
   int Stat(const NlcUrl& url, struct __stat64* buffer) override;
 
-  ssize_t Read(void* lpBuf, size_t uiBufSize) override;
+  int64_t Read(void* lpBuf, size_t uiBufSize) override;
   int64_t Seek(int64_t iFilePosition, int iWhence = SEEK_SET) override;
   void Close() override;
   int64_t GetPosition() override;
