+++ nolimitlib/DependLibs/libplayer-nlc/xbmc/xbmc/filesystem/OverrideFile.h
@@ -27,8 +27,8 @@ public:
   bool Delete(const NlcUrl& url) override;
   bool Rename(const NlcUrl& url, const NlcUrl& urlnew) override;
 
-  ssize_t Read(void* lpBuf, size_t uiBufSize) override;
-  ssize_t Write(const void* lpBuf, size_t uiBufSize) override;
+  int64_t Read(void* lpBuf, size_t uiBufSize) override;
+  int64_t Write(const void* lpBuf, size_t uiBufSize) override;
   int64_t Seek(int64_t iFilePosition, int iWhence = SEEK_SET) override;
   void Close() override;
   int64_t GetPosition() override;
