+++ nolimitlib/DependLibs/libplayer-nlc/xbmc/xbmc/filesystem/PipeFile.h
@@ -39,8 +39,8 @@ public:
   bool Exists(const NlcUrl& url) override;
   int Stat(const NlcUrl& url, struct __stat64* buffer) override;
   int Stat(struct __stat64* buffer) override;
-  ssize_t Read(void* lpBuf, size_t uiBufSize) override;
-  ssize_t Write(const void* lpBuf, size_t uiBufSize) override;
+  int64_t Read(void* lpBuf, size_t uiBufSize) override;
+  int64_t Write(const void* lpBuf, size_t uiBufSize) override;
   int64_t Seek(int64_t iFilePosition, int iWhence = SEEK_SET) override;
   void Close() override;
   void Flush() override;
