+++ nolimitlib/DependLibs/libplayer-nlc/xbmc/xbmc/filesystem/ZipFile.h
@@ -27,7 +27,7 @@ namespace XFILE
     bool Exists(const NlcUrl& url) override;
     int Stat(struct __stat64* buffer) override;
     int Stat(const NlcUrl& url, struct __stat64* buffer) override;
-    ssize_t Read(void* lpBuf, size_t uiBufSize) override;
+    int64_t Read(void* lpBuf, size_t uiBufSize) override;
     //virtual bool ReadString(char *szLine, int iLineLength);
     int64_t Seek(int64_t iFilePosition, int iWhence = SEEK_SET) override;
     void Close() override;
