+++ nolimitlib/DependLibs/libplayer-nlc/xbmc/xbmc/platform/posix/filesystem/PosixFile.h
@@ -22,8 +22,8 @@ namespace XFILE
     bool OpenForWrite(const NlcUrl& url, bool bOverWrite = false) override;
     void Close() override;
     
-    ssize_t Read(void* lpBuf, size_t uiBufSize) override;
-    ssize_t Write(const void* lpBuf, size_t uiBufSize) override;
+    int64_t Read(void* lpBuf, size_t uiBufSize) override;
+    int64_t Write(const void* lpBuf, size_t uiBufSize) override;
     int64_t Seek(int64_t iFilePosition, int iWhence = SEEK_SET) override;
     int Truncate(int64_t size) override;
     int64_t GetPosition() override;
